refactor: modularize server routing
